+++ library/src/main/java/org/dbtools/android/commons/ui/compose/form/DropdownMenuBoxField.kt
@@ -13,6 +13,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import org.dbtools.android.commons.ui.compose.DayNightTextField
 
@@ -23,9 +24,18 @@ fun <T> DropdownMenuBoxField(
     onOptionSelected: (T) -> Unit,
     optionToText: @Composable (T) -> String,
     modifier: Modifier = Modifier,
-    label: String? = null
+    label: String? = null,
+    supportingText: String? = null,
+    errorTextFlow: StateFlow<String?> = MutableStateFlow(null),
 ) {
     val selectedOption by selectedOptionFlow.collectAsStateWithLifecycle()
+    val errorText by errorTextFlow.collectAsStateWithLifecycle()
+
+    val supportText = when {
+        !errorText.isNullOrBlank() -> errorText
+        !supportingText.isNullOrBlank() -> supportingText
+        else -> null
+    }
 
     DropdownMenuBoxField(
         options = options,
@@ -33,7 +43,9 @@ fun <T> DropdownMenuBoxField(
         onOptionSelected = onOptionSelected,
         optionToText = optionToText,
         modifier = modifier,
-        label = label
+        label = label,
+        supportingText = supportText,
+        isError = !errorText.isNullOrBlank()
     )
 }
 
@@ -44,7 +56,9 @@ fun <T> DropdownMenuBoxField(
     onOptionSelected: (T) -> Unit,
     optionToText: @Composable (T) -> String,
     modifier: Modifier = Modifier,
-    label: String? = null
+    label: String? = null,
+    supportingText: String? = null,
+    isError: Boolean = false
 ) {
     var expanded by remember { mutableStateOf(false) }
 
@@ -58,6 +72,8 @@ fun <T> DropdownMenuBoxField(
             onValueChange = {},
             label = if (label != null) { { Text(text = label) } } else null,
             trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) },
+            supportingText = supportingText?.let{ { Text(it) } },
+            isError = isError,
             modifier = modifier
                 // As of Material3 1.0.0-beta03; The `menuAnchor` modifier must be passed to the text field for correctness.
                 // (https://android-review.googlesource.com/c/platform/frameworks/support/+/2200861)