+++ library/src/main/java/org/dbtools/android/commons/ui/compose/form/TextFieldData.kt
@@ -1,5 +1,6 @@
 package org.dbtools.android.commons.ui.compose.form
 
+@Deprecated("Use FlowTextField")
 data class TextFieldData(
     val text: String,
     val supportingText: String? = null,